[vllm] Simplify Snowball parity test structure
Consolidate representative serving parity under the Snowball model name, isolate checkpoint-only dependencies, and remove redundant short-prompt and fixture tests.
