--- tests/cluster/vllm/resources/june_tpu_67b_a2b_step_42150_logprobs.json
@@ -1,34 +1,0 @@
-{
-  "moe_implementation": "sonic",
-  "mp": "params=float32,compute=bfloat16,output=bfloat16",
-  "prompt": "The United States Of",
-  "prompt_token_ids": [791, 3723, 4273, 5046],
-  "tokenizer": "marin-community/marin-tokenizer",
-  "top_logprobs": [
-    {"logprob": -0.4749794006347656, "text": " America", "token_id": 5270},
-    {"logprob": -4.453495025634766, "text": " The", "token_id": 578},
-    {"logprob": -4.734745025634766, "text": " Tara", "token_id": 70797},
-    {"logprob": -4.765995025634766, "text": " Africa", "token_id": 10384},
-    {"logprob": -5.015995025634766, "text": " Am", "token_id": 3383},
-    {"logprob": -5.265995025634766, "text": " Europe", "token_id": 4606},
-    {"logprob": -5.297245025634766, "text": " A", "token_id": 362},
-    {"logprob": -5.297245025634766, "text": " Americ", "token_id": 2340},
-    {"logprob": -5.422245025634766, "text": " Beer", "token_id": 34484},
-    {"logprob": -5.578495025634766, "text": " Ar", "token_id": 1676},
-    {"logprob": -5.609745025634766, "text": " India", "token_id": 6890},
-    {"logprob": -5.609745025634766, "text": " Football", "token_id": 21424},
-    {"logprob": -5.640995025634766, "text": " Marijuana", "token_id": 65939},
-    {"logprob": -5.734745025634766, "text": " Murder", "token_id": 52483},
-    {"logprob": -5.859745025634766, "text": " Food", "token_id": 12369},
-    {"logprob": -5.890995025634766, "text": " Russia", "token_id": 8524},
-    {"logprob": -5.922245025634766, "text": " Sports", "token_id": 13482},
-    {"logprob": -6.015995025634766, "text": "\n", "token_id": 198},
-    {"logprob": -6.015995025634766, "text": " North", "token_id": 4892},
-    {"logprob": -6.015995025634766, "text": " Soccer", "token_id": 38633},
-    {"logprob": -6.047245025634766, "text": " E", "token_id": 469},
-    {"logprob": -6.140995025634766, "text": " Canada", "token_id": 7008},
-    {"logprob": -6.203495025634766, "text": " Hate", "token_id": 66912},
-    {"logprob": -6.234745025634766, "text": " American", "token_id": 3778},
-    {"logprob": -6.265995025634766, "text": " Mind", "token_id": 23754}
-  ]
-}
